requirements.txt
pip
Flask
flask_table
requests
cassandra-driver
wtforms
passlib
